Bonus Homework Class2 Upload

+++ Class 2/SolutionClass2/AverageNumber/Program.cs
@@ -18,14 +18,14 @@ namespace AverageNumber
             Console.WriteLine("Enter the fourth number:");
             string fourthNumberString = Console.ReadLine();
 
-            bool successfulConversion1 = int.TryParse(firstNumberString, out int firstNumber);
-            bool successfulConversion2 = int.TryParse(secondNumberString, out int secondNumber);
-            bool successfulConversion3 = int.TryParse(thirdNumberString, out int thirdNumber);
-            bool successfulConversion4 = int.TryParse(fourthNumberString, out int fourthNumber);
+            bool successfulConversion1 = float.TryParse(firstNumberString, out float firstNumber);
+            bool successfulConversion2 = float.TryParse(secondNumberString, out float secondNumber);
+            bool successfulConversion3 = float.TryParse(thirdNumberString, out float thirdNumber);
+            bool successfulConversion4 = float.TryParse(fourthNumberString, out float fourthNumber);
 
             if(successfulConversion1 && successfulConversion2 && successfulConversion3 && successfulConversion4)
             {
-                int result = (firstNumber + secondNumber + thirdNumber + fourthNumber) / 4;
+                float result = (firstNumber + secondNumber + thirdNumber + fourthNumber) / 4;
                 Console.WriteLine("The average of " + firstNumber + ", " + secondNumber + ", " + thirdNumber + " and " + fourthNumber + " is: " + result);
             }
             else
